# doc/openapi.yaml
openapi: 3.0.0
info:
  title: 飲食店向け簡易オーダーシステム
  version: "1.0"
servers:
  - url: "http://localhost:3000"
...
paths:
...
  /management/order/byRestaurant/{restaurant_id}:
    get:
      tags:
        - management
      summary: 指定した店舗の注文一覧の取得
      operationId: orderList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: 注文ID
                        ordered_date:
                          type: string
                          description: 注文日時
                        seat_id:
                          type: integer
                          description: 座席ID
                        seat_name:
                          type: string
                          description: 座席名
                        total_price:
                          type: integer
                          description: 合計金額
                        is_all_delivered:
                          type: boolean
                          description: 全ての料理が配達済みかどうか
                        is_order_finished:
                          type: boolean
                          description: 注文が完了しているかどうか
                        is_paid:
                          type: boolean
                          description: 支払いが完了しているかどうか
        "400":
          description: Bad Request (restaurant_idに不備がある)
        "401":
          description: Unauthorized (トークンが無効)
        "403":
          description: Permission Denied (owner, kitchen, counterは自分の店舗のみ取得可能)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: authorization
          description: /management/login で取得したトークン
        - schema:
            type: integer
          in: path
          required: true
          name: restaurant_id
          description: 注文一覧を取得したい店舗のID
      security:
        - bearerAuth: []
  /management/order/byRestaurant/{restaurant_id}/unserved:
    get:
      tags:
        - management
      summary: 指定した店舗の未提供料理一覧を取得 (注文時間順)
      operationId: unservedDishList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  unserved_dishes:
                    type: array
                    items:
                      type: object
                      properties:
                        order_id:
                          type: integer
                          description: 注文ID
                        ordered_dish_id:
                          type: integer
                          description: 注文した料理のID
                        dish_id:
                          type: integer
                          description: 注文した料理の料理ID
                        dish_name:
                          type: string
                          description: 注文した料理の名前
                        quantity:
                          type: integer
                          description: 注文した料理の数量
                        seat_id:
                          type: integer
                          description: 座席ID
                        seat_name:
                          type: string
                          description: 座席名
                        ordered_date:
                          type: string
                          description: 注文日時
        "400":
          description: Bad Request (restaurant_idに不備がある)
        "401":
          description: Unauthorized (トークンが無効)
        "403":
          description: Permission Denied (owner, counter, kitchenは自分の店舗のみ取得可能)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: authorization
          description: /management/login で取得したトークン
        - schema:
            type: integer
          in: path
          required: true
          name: restaurant_id
          description: 未提供料理一覧を取得したい店舗のID
      security:
        - bearerAuth: []
  /management/order/{order_id}:
    get:
      tags:
        - management
      summary: 注文詳細の取得
      operationId: orderGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id:
                    type: integer
                    description: 注文ID
                  ordered_date:
                    type: string
                    description: 注文日時
                  seat_id:
                    type: integer
                    description: 座席ID
                  seat_name:
                    type: string
                    description: 座席名
                  is_all_delivered:
                    type: boolean
                    description: 全ての料理が配達済みかどうか
                  is_order_finished:
                    type: boolean
                    description: 注文が完了しているかどうか
                  is_paid:
                    type: boolean
                    description: 支払いが完了しているかどうか
                  ordered_items:
                    type: array
                    items:
                      type: object
                      properties:
                        ordered_dish_id:
                          type: integer
                          description: 注文した料理のID
                        dish_id:
                          type: integer
                          description: 注文した料理の料理ID
                        dish_name:
                          type: string
                          description: 注文した料理の名前
                        dish_price:
                          type: integer
                          description: 注文した料理の単価
                        dish_quantity:
                          type: integer
                          description: 注文した料理の数量
                        is_delivered:
                          type: boolean
                          description: その料理が配達済みかどうか
                  total_price:
                    type: integer
                    description: 合計金額
        "400":
          description: Bad Request (order_idに不備がある)
        "401":
          description: Unauthorized (トークンが無効)
        "403":
          description: Permission Denied (owner, kitchen, counterは自分の店舗のみ取得可能)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: authorization
          description: /management/login で取得したトークン
        - schema:
            type: integer
          in: path
          required: true
          name: order_id
          description: 注文詳細を取得したい注文のID
      security:
        - bearerAuth: []
...
  /management/order/byOrderedDish/{ordered_dish_id}/cancel:
    put:
      tags:
        - management
      summary: 注文した(+未提供)料理のキャンセル
      operationId: orderedDishCancel
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request (ordered_dish_idに不備がある、もしくは提供済みでキャンセルできない)
        "401":
          description: Unauthorized (トークンが無効)
        "403":
          description: Permission Denied (owner, counterは自分の店舗の注文のみキャンセル可能、kitchenはアクセス不可)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: authorization
          description: /management/login で取得したトークン
        - schema:
            type: integer
          in: path
          required: true
          name: ordered_dish_id
          description: キャンセルしたい料理の注文ID
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT